fixed compile warnings on 64 bits
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3332 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/carrierroute/carrier_tree.c
@@ -289,7 +289,9 @@ struct carrier_tree * add_carrier_tree(const char * carrier, int carrier_id, str
 		return NULL;
 	}
 	rd->carriers[id]->index = id;
-	LM_INFO("created carrier tree: %s, with id %i and %i trees\n", rd->carriers[id]->name.s, rd->carriers[id]->id, rd->carriers[id]->tree_num);
+	LM_INFO("created carrier tree: %s, with id %i and %ld trees\n", 
+		rd->carriers[id]->name.s, rd->carriers[id]->id, 
+		(long)rd->carriers[id]->tree_num);
 	return rd->carriers[id];
 }
